libm: Add optimized expf
This implementation is based on code from ARM optimized routines [1].
The error for expf, for nearest rouding mode, is estimated on 0.502 ulp
with worst case in non-nearest up to 1 ulp.

Double precision arithmetics is used with internal calculation and it
is expected to be faster and easier to get better precision on most
targets.  The const data is kept in a separate translation unit that
allows fast code for literal loads on most targets and also allows
sharing data across expf, exp2f and powf.

The performance shows about 2x better throughput and 1.8x better
latency for aarch64 (A72) using expf benchmark.

Some details may need target specific tweaks:

 * Best convert and round to int operation in the arg reduction may be
   different across targets.  This is defined by two preprocessors:

   - HAVE_FAST_ROUND: which indicates the target have hardware
     instruction for the round operation.

   - HAVE_FAST_LROUND: as before, but for lround operation.

   Currently aarch64 is defined to have support for both.

 * An extra configuration file is added for bionic (bionic.h) to correct
   set the library expectations regarding rounding and errno setting.

 * Code was optimized for targets with FMA instruction (optimal
   polynomial eval may be different without fma).

 * Compiler might not always generate good code for floating bit
   representation access via unions or it may be inherently slow on some
   targets.

[1] https://github.com/ARM-software/optimized-routines

Test: ran 32-bit and 64-bit x86 tests on host
Change-Id: If67293b450e32666c4d40482ea0380f661ad5581
Signed-off-by: mydongistiny <jaysonedson@gmail.com>